Moving fetchJobSummary, fetchJobSummary API calls to ExtractorStatus component to fix interval issue
