+++ frontend/src/components/listeners/SubmitExtraction.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from "react";
+import React, {useState} from "react";
 import {useDispatch, useSelector} from "react-redux";
 import {
 	Box,
@@ -24,8 +24,6 @@ import {Extractor, RootState} from "../../types/data";
 import {ClowderRjsfSelectWidget} from "../styledComponents/ClowderRjsfSelectWidget";
 import {ClowderRjsfTextWidget} from "../styledComponents/ClowderRjsfTextWidget";
 import ExtractorStatus from "./ExtractorStatus";
-import { fetchJobSummary, fetchJobUpdates } from "../../actions/listeners";
-
 
 type SubmitExtractionProps = {
 	fileId: string,
@@ -50,24 +48,15 @@ export default function SubmitExtraction(props: SubmitExtractionProps) {
 	const submitDatasetExtraction =
 		(datasetId: string | undefined, extractorName: string | undefined, requestBody: FormData) => dispatch(submitDatasetExtractionAction(datasetId, extractorName, requestBody));
    
-	const jobSummaryFetch = (job_id: string | undefined) => dispatch(fetchJobSummary(job_id));
-	const jobUpdatesFetch = (job_id: string | undefined) => dispatch(fetchJobUpdates(job_id));
-    
 	const job_id = useSelector((state: RootState) => state.listener.currJobId);
-	const summary = useSelector((state: RootState) => state.listener.currJobSummary);
-	const updates = useSelector((state: RootState) => state.listener.currJobUpdates);
 
-	const [isFormSubmitted, setIsFormSubmitted] = useState(false);
-
-    const onSubmit = (formData: FormData) => {
+	const onSubmit = (formData: FormData) => {
 		const extractorName = selectedExtractor.name
 		if (fileId === undefined && datasetId !== undefined) {
 			submitDatasetExtraction(datasetId, extractorName, formData);
-			setIsFormSubmitted(true);
 			handleNext();
 		} else if (fileId !== undefined) {
 			submitFileExtraction(fileId, extractorName, formData);
-			setIsFormSubmitted(true);
 			handleNext();
 		}
 	}
@@ -86,18 +75,6 @@ export default function SubmitExtraction(props: SubmitExtractionProps) {
 		handleClose();
 		setActiveStep(0);
 	}
-
-	useEffect(() => {
-		const interval = setInterval(() => {
-			if (isFormSubmitted == true && job_id.length > 0) {
-				jobSummaryFetch(job_id);
-				jobUpdatesFetch(job_id);
-			}
-		}, 2000);
-
-		return () => clearInterval(interval);
-	});
-
 
 	return (
 		// TODO replace this with submit extraction content
@@ -156,11 +133,7 @@ export default function SubmitExtraction(props: SubmitExtractionProps) {
 						<Step key="status">
 							<StepLabel>Extraction Status</StepLabel>
 							<StepContent>
-								<ExtractorStatus 
-									job_id={job_id}
-									summary={summary}
-									updates={updates}
-								/>
+								<ExtractorStatus job_id={job_id}/>
 								{/*buttons*/}
 								<Box sx={{mb: 2}}>
 									<Button variant="contained" onClick={handleNext} sx={{mt: 1, mr: 1}}>
